requirements.txt
langchain
langgraph
google-generativeai
python-dotenv
mcp
requests
uv